refactor: reorganize authentication modules and decorators into a shared directory; remove unused ZodValidationPipe
