+++ src/domain/_shared/auth/auth.module.ts
@@ -5,10 +5,10 @@ import { PassportModule } from '@nestjs/passport';
 import { APP_GUARD } from '@nestjs/core';
 import { Env } from '@/shared/env/env';
 import { PrismaService } from '@/shared/prisma/prisma.service';
-import { JwtStrategy } from './jwt.strategy';
-import { AuthenticateModule } from './authenticate/authenticate.module';
-import { JwtAuthGuard } from './jwt-auth.guard';
-import { RolesGuard } from './roles/roles.guard';
+import { JwtStrategy } from '../auth/jwt/jwt.strategy';
+import { AuthenticateModule } from '../../auth/authenticate/authenticate.module';
+import { JwtAuthGuard } from './jwt/jwt-auth.guard';
+import { RolesGuard } from '@/domain/_shared/auth/roles/roles.guard';
 
 @Global()
 @Module({
