+++ src/domain/_shared/auth/decorators/current-user.decorator.ts
@@ -1,5 +1,5 @@
 import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-import { UserPayload } from './jwt.strategy';
+import { UserPayload } from '../jwt/jwt.strategy';
 
 export const CurrentUser = createParamDecorator(
   (_: never, context: ExecutionContext) => {
