+++ src/domain/_shared/auth/jwt/jwt.strategy.ts
