+++ src/domain/_shared/auth/roles/roles.guard.ts
@@ -7,7 +7,7 @@ import {
 } from '@nestjs/common';
 import { Reflector } from '@nestjs/core';
 import { Role } from '@prisma/client';
-import { ROLES_KEY } from './roles.decorator';
+import { ROLES_KEY } from '../../_shared/auth/decorators/roles.decorator';
 
 @Injectable()
 export class RolesGuard implements CanActivate {
