+++ src/domain/identity/admin/update-user-to-moderator/update-user-to-moderator.controller.ts
@@ -2,8 +2,8 @@ import {
   Controller, Param, Patch, UseGuards,
 } from '@nestjs/common';
 import { UpdateUserToModeratorService } from './update-user-to-moderator.service';
-import { JwtAuthGuard } from '@/domain/auth/jwt-auth.guard';
-import { RolesGuard } from '@/domain/auth/roles/roles.guard';
+import { JwtAuthGuard } from '@/domain/_shared/auth/jwt/jwt-auth.guard';
+import { RolesGuard } from '@/domain/_shared/auth/roles/roles.guard';
 import { UpdateUserToModeratorDto } from './update-user-to-moderator.dto';
 
 @Controller('admin/:userId/add-moderator-role')
