+++ src/domain/identity/user/disable-user/disable-user.controller.ts
@@ -1,10 +1,10 @@
 import {
   Controller, UseGuards, Patch, Param,
 } from '@nestjs/common';
-import { JwtAuthGuard } from '@/domain/auth/jwt-auth.guard';
-import { RolesGuard } from '@/domain/auth/roles/roles.guard';
-import { CurrentUser } from '@/domain/auth/current-user.decorator';
-import { UserPayload } from '@/domain/auth/jwt.strategy';
+import { JwtAuthGuard } from '@/domain/_shared/auth/jwt/jwt-auth.guard';
+import { RolesGuard } from '@/domain/_shared/auth/roles/roles.guard';
+import { CurrentUser } from '@/domain/_shared/auth/decorators/current-user.decorator';
+import { UserPayload } from '@/domain/_shared/auth/jwt/jwt.strategy';
 import { DisableUserService } from './disable-user.service';
 import { DisableUserDto } from './disable-user.dto';
