+++ src/domain/identity/user/disable-user/disable-user.service.ts
@@ -1,7 +1,7 @@
 import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { Role } from '@prisma/client';
 import { PrismaService } from '@/shared/prisma/prisma.service';
-import { UserPayload } from '@/domain/auth/jwt.strategy';
+import { UserPayload } from '@/domain/_shared/auth/jwt/jwt.strategy';
 
 @Injectable()
 export class DisableUserService {
