+++ src/domain/identity/user/register-with-artisan/register-with-artisan.controller.ts
@@ -1,7 +1,7 @@
 import { Body, Controller, Post } from '@nestjs/common';
 import { RegisterWithArtisanService } from './register-with-artisan.service';
 import { RegisterWithArtisanDto } from './register-with-artisan.dto';
-import { Public } from '@/domain/auth/public.decorator';
+import { Public } from '@/domain/_shared/auth/decorators/public.decorator';
 
 @Controller('users/register-with-artisan')
 export class RegisterWithArtisanController {
