+++ src/domain/test/test.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { APP_GUARD } from '@nestjs/core';
 import { TestController } from './test.controller';
 import { TestService } from './test.service';
-import { RolesGuard } from '@/domain/auth/roles/roles.guard';
+import { RolesGuard } from '@/domain/_shared/auth/roles/roles.guard';
 
 @Module({
   controllers: [TestController],
