--- src/shared/pipes/zod-validation.pipe.ts
@@ -1,24 +1,0 @@
-import { PipeTransform, BadRequestException } from '@nestjs/common';
-import { ZodError, ZodSchema } from 'zod';
-import { fromZodError } from 'zod-validation-error';
-
-export class ZodValidationPipe implements PipeTransform {
-  constructor(private schema: ZodSchema) {}
-
-  transform(value: unknown) {
-    try {
-      const parsedValue = this.schema.parse(value);
-      return parsedValue;
-    } catch (error) {
-      if (error instanceof ZodError) {
-        throw new BadRequestException({
-          message: 'Validation failed',
-          statusCode: 400,
-          errors: fromZodError(error),
-        });
-      }
-
-      throw new BadRequestException('Validation failed');
-    }
-  }
-}
